@@ -1,36 +1,26 @@
-name: imagebutton_example
-description: Demonstrates how to use the image_button plugin.
-publish_to: 'none'
+name            : imagebutton_example
+description     : Demonstrates how to use the image_button plugin.
+publish_to      : 'none'
 
-environment:
-  sdk: ">=2.1.0 <3.0.0"
+environment     :
+  sdk: '>=3.2.6 <4.0.0'
 
-dependencies:
-  flutter:
+dependencies    :
+  flutter        :
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  imagebutton    :
+    path: ../
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
-  flutter_test:
+  integration_test:
     sdk: flutter
+  flutter_test    :
+    sdk: flutter
+  flutter_lints   : ^2.0.0
 
-  imagebutton:
-    path: ../
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
-flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
+flutter         :
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
-  assets:
+  assets              :
     - assets/
