@@ -1,14 +1,19 @@
 import Flutter
 import UIKit
 
-public class SwiftImagebuttonPlugin: NSObject, FlutterPlugin {
+public class ImagebuttonPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "imagebutton", binaryMessenger: registrar.messenger())
-    let instance = SwiftImagebuttonPlugin()
+    let instance = ImagebuttonPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    result("iOS " + UIDevice.current.systemVersion)
+    switch call.method {
+    case "getPlatformVersion":
+      result("iOS " + UIDevice.current.systemVersion)
+    default:
+      result(FlutterMethodNotImplemented)
+    }
   }
 }
